Big refactor and tidy up
