--- extensions/plexamp/src/search-albums.tsx
@@ -1,5 +1,0 @@
-import { SearchCommand } from "./search-shared";
-
-export default function Command() {
-  return <SearchCommand />;
-}
