+++ extensions/plexamp/src/use-library-selection.ts
@@ -1,58 +1,42 @@
-import { useCallback, useEffect, useState } from "react";
+import { useCallback } from "react";
 
 import {
   getMusicSections,
   getPlexSetupStatus,
   resolveSelectedLibrary,
 } from "./plex";
+import { useAsyncValue } from "./use-async-value";
 import type { LibrarySection } from "./types";
 
-interface LibrarySelectionState {
-  isLoading: boolean;
-  libraries: LibrarySection[];
-  selectedLibrary?: LibrarySection;
-  selectedServerName?: string;
-  error?: string;
-}
-
 export function useLibrarySelection() {
-  const [state, setState] = useState<LibrarySelectionState>({
-    isLoading: true,
-    libraries: [],
-  });
-
-  const reload = useCallback(async () => {
-    setState((current) => ({ ...current, isLoading: true, error: undefined }));
-
-    try {
+  const loadState = useAsyncValue(
+    useCallback(async () => {
       const [libraries, setupStatus] = await Promise.all([
         getMusicSections(),
         getPlexSetupStatus(),
       ]);
       const selectedLibrary = await resolveSelectedLibrary(libraries);
-      setState({
-        isLoading: false,
+
+      return {
         libraries,
         selectedLibrary,
         selectedServerName: setupStatus.selectedServerName,
-      });
-    } catch (error) {
-      setState({
-        isLoading: false,
-        libraries: [],
-        selectedLibrary: undefined,
-        selectedServerName: undefined,
-        error: error instanceof Error ? error.message : String(error),
-      });
-    }
-  }, []);
-
-  useEffect(() => {
-    void reload();
-  }, [reload]);
+      };
+    }, []),
+    "library-selection",
+    {
+      libraries: [] as LibrarySection[],
+      selectedLibrary: undefined,
+      selectedServerName: undefined,
+    },
+  );
 
   return {
-    ...state,
-    reload,
+    isLoading: loadState.isLoading,
+    libraries: loadState.value.libraries,
+    selectedLibrary: loadState.value.selectedLibrary,
+    selectedServerName: loadState.value.selectedServerName,
+    error: loadState.error,
+    reload: loadState.reload,
   };
 }
